backend/requirements.txt
Flask==3.0.3
meilisearch==0.31.5
flask-cors==3.0.10
numpy==2.1.1
pandas==2.2.3
python-dotenv==1.0.1